+++ app/controllers/home_controller.rb
@@ -2,9 +2,9 @@ class HomeController < ApplicationController
   def index
     
     # Featured Projects
-    @featured_projects = Project.limit(3) # Adjust as necessary for your featured logic
+    @featured_projects = Project.where(featured: true)
     @featured_blog_posts = BlogPost.where(featured: true)
-    @featured_videos = Video.limit(3) # Adjust as necessary for your featured logic
+    @featured_videos = Video.where(featured: true)
     
     #