+++ app/controllers/videos_controller.rb
@@ -1,10 +1,9 @@
 class VideosController < ApplicationController
   def index
-    @featured_videos = Video.limit(3) # Adjust as necessary for your featured logic
     @videos = Video.all
   end
 
   def show
-    @featured_video = Video.find(params[:id])
+    @video = Video.find(params[:id])
   end
 end