+++ db/migrate/20241019162525_add_featured_to_videos.rb
@@ -1,0 +1,5 @@
+class AddFeaturedToVideos < ActiveRecord::Migration[7.1]
+  def change
+    add_column :videos, :featured, :boolean
+  end
+end