+++ db/migrate/20241019162534_add_featured_to_projects.rb
@@ -1,0 +1,5 @@
+class AddFeaturedToProjects < ActiveRecord::Migration[7.1]
+  def change
+    add_column :projects, :featured, :boolean
+  end
+end