+++ db/seeds/videos.rb
@@ -1,17 +1,20 @@
 Video.create!(
   title: "Pickled Pirates at Nostalgia Drags 2024",
   description: "What a fantastic way to celebrate my birthday weekend! The Nostalgia Drags at Toutle Motorsports Park brought together some of my favorite things—friends gathering to enjoy the thrill of 'old school' drag racing. The event featured an incredible car show, mouthwatering BBQ, and even a scenic river to relax by when we needed to cool off. It was a perfect blend of adrenaline and camaraderie, creating unforgettable memories with good company. I couldn’t have asked for a better way to mark the occasion!",
-  youtube_id: "T5fi5Bkn5qk?si=bUJDUXuGRnWHW9Fk" # See readme
+  youtube_id: "T5fi5Bkn5qk?si=bUJDUXuGRnWHW9Fk", # See readme
+  featured: true
 )
 
 Video.create!(
   title: "Tahuya Off-Road 2024",
   description: "Another fantastic weekend of camping, 4x4 off-roading, and everything in between! I couldn’t have asked for a better group of people to share the experience with. The highlight was definitely taking Ryder on his first real camping adventure; seeing his excitement made it all the more special. Plus, I can’t forget to shout out a happy birthday to one of my best friends, Daniel! It was an amazing way to celebrate and create lasting memories together in the great outdoors.",
-  youtube_id: "EVho5M4SfGY?si=pD2I_0G-ganKWZ3e" # See readme
+  youtube_id: "EVho5M4SfGY?si=pD2I_0G-ganKWZ3e", # See readme
+  featured: true
 )
 
 Video.create!(
   title: "WJ Cherokee - 'Break' In",
   description: "This is one of the first fabrication projects I completed in my shop! This WJ Cherokee features a 4-link suspension setup, Dana axle swaps on both ends, coilover suspension up front, and air shocks in the rear, resulting in an impressive amount of flex. It was an exciting project that allowed me to apply my skills and creativity in automotive fabrication. Check out the video to see the finished product in action and get a closer look at the features that make this build stand out!",
-  youtube_id: "Zrfsuu33IYI?si=6K_RRyYlqlLxGjuA" # See readme
+  youtube_id: "Zrfsuu33IYI?si=6K_RRyYlqlLxGjuA", # See readme
+  featured: true
 )