+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] 2024/10/19
+## Other 'featured' migrations
+
+## Added
+- Corresponding schema and migration files related migrations run
+
+## Changed
+- home_controller.rb
+    - Featured logic was actually here for the home page related to featured projects and videos. Updated that.
+- seeds.rb
+    - Updated to add 'featured' data
+
 ## [1.3.0] 2024/10/19
 ## Migration for better featured blog post control, database correction